Repair install-playstore.sh
I, and many other users, faced the problem that Google Play is not installed, the script completes without error, after downloading gapps, but Google Play does not appear in the list of applications. I managed to solve the issue with installing google play. The matter was that the gapps package was not downloaded, I downloaded it manually, corrected the script (IIt is better to check the one who understands this), now, if there is this gapps archive (in my version of the script the following name of the file gapps is written: "open_gapps-x86_64-7.1-pico-20211103.zip") in the home directory, the script works.